requirements.txt
selenium
webdriver-manager
beautifulsoup4
undetected-chromedriver
fastapi
uvicorn[standard]
python-dotenv
pydantic
